separate big-picture.js on modules
